@@ -29,35 +47,71 @@
     "GIS"
   ],
   "author": "Kohei Otsuka, Code for History",
-  "license": "Apache 2.0",
+  "license": "Maplat Limited License 1.1",
   "bugs": {
     "url": "https://github.com/code4history/MaplatCore/issues"
   },
   "homepage": "https://github.com/code4history/MaplatCore/wiki",
   "dependencies": {
-    "@turf/turf": "^5.1.6",
-    "i18next": "^17.0.14",
-    "i18next-xhr-backend": "^3.1.2",
-    "@maplat/tin": "^0.6.3",
-    "ol": "^5.3.3"
+    "@maplat/tin": "^0.9.0",
+    "@turf/line-intersect": "^6.5.0",
+    "@turf/turf": "^6.5.0",
+    "argv": "0.0.2",
+    "i18next": "^17.3.1",
+    "i18next-xhr-backend": "^3.2.2",
+    "lodash.template": "^4.5.0",
+    "ol": "^6.1.1",
+    "weiwudi": "^0.1.4",
+    "whatwg-fetch": "^3.6.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.5.5",
-    "@babel/core": "^7.5.5",
-    "@babel/polyfill": "^7.4.4",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/register": "^7.5.5",
-    "babel-loader": "^8.0.6",
-    "core-js": "^3.2.1",
-    "eslint": "^6.3.0",
+    "@babel/cli": "^7.14.5",
+    "@babel/core": "^7.14.6",
+    "@babel/plugin-proposal-class-properties": "^7.14.5",
+    "@babel/preset-env": "^7.14.7",
+    "@babel/preset-typescript": "^7.14.5",
+    "@types/jest": "^26.0.23",
+    "@types/lodash.template": "^4.5.0",
+    "@types/node": "^14.17.4",
+    "@types/ol": "^6.5.1",
+    "@types/eslint": "^7.2.13",
+    "@types/estree": "0.0.48",
+    "@typescript-eslint/eslint-plugin": "^4.28.0",
+    "@typescript-eslint/parser": "^4.28.0",
+    "babel-jest": "^26.6.3",
+    "babel-loader": "^8.2.2",
+    "clean-webpack-plugin": "^3.0.0",
+    "core-js": "^3.15.1",
+    "css-loader": "^5.2.6",
+    "css-minimizer-webpack-plugin": "^1.3.0",
+    "eslint": "^7.29.0",
+    "eslint-config-prettier": "^7.2.0",
+    "eslint-loader": "^4.0.2",
+    "eslint-plugin-jest": "^24.3.6",
+    "file-loader": "^6.2.0",
+    "fork-ts-checker-webpack-plugin": "^6.2.12",
     "fs-extra": "^8.1.0",
     "gulp": "^4.0.2",
-    "gulp-cli": "^2.2.0",
+    "gulp-cli": "^2.3.0",
     "gulp-zip": "^4.2.0",
-    "jasmine": "^3.4.0",
-    "less": "^3.10.3",
-    "less-plugin-clean-css": "^1.5.1",
-    "webpack": "^4.39.3",
-    "webpack-cli": "^3.3.7"
+    "html-webpack-plugin": "^4.5.2",
+    "jest": "^26.6.3",
+    "jest-cli": "^26.6.3",
+    "jest-puppeteer": "^4.4.0",
+    "less": "^3.13.1",
+    "less-loader": "^7.3.0",
+    "mini-css-extract-plugin": "^1.6.0",
+    "prettier": "^2.3.1",
+    "psaux": "^0.4.0",
+    "puppeteer": "^5.5.0",
+    "sleep-promise": "^9.1.0",
+    "terminate": "^2.2.0",
+    "terser-webpack-plugin": "^4.2.3",
+    "ts-jest": "^26.5.6",
+    "typescript": "^4.3.4",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.40.0",
+    "webpack-cli": "^4.7.2",
+    "webpack-dev-server": "^3.11.2"
   }
 }
